network connection

+++ mzapo_template/lcdframe.h
@@ -17,7 +17,7 @@ void clear_screen();
 void frameToLCD();
 int charToFrame(char c, int yRow, int xColumn, uint16_t forecolor, uint16_t backcolor);
 int strToFrame(char *str, int yRow, int xColumn, uint16_t forecolor, uint16_t backcolor, bool big);
-int int_to_frame(int number, int yRow, int xColumn, uint16_t forecolor, uint16_t backcolor, bool big); 
+int int_to_frame(unsigned long number, int yRow, int xColumn, uint16_t forecolor, uint16_t backcolor, bool big);
 
 #ifdef __cplusplus
 } /* extern "C" */
